# swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Altzone API",
        "version": "1.0.0",
        "description": "This is a description of API for the Altzone game.\nThis swagger document is describing endpoint in more details and includes the request objects structure, examples, authorization rules and any other additional information.\nNotice,  that it does not include common rules and functions, since they are covered on the wikipages in GitHub",
        "x-logo": {
            "url": ""
        }
    },
    "servers": [
        {
            "url": "https://altzone.fi/api",
            "description": "Production. Stable version of API, which might not have all of the new changes"
        },
        {
            "url": "http://localhost:8080",
            "description": "Local"
        },
        {
            "url": "https://devapi.altzone.fi",
            "description": "Development. Current version of API with latest changes"
        }
    ],
    "paths": {
...
        "/chat": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200"
                                },
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Chat": [
                                                    {
                                                        "objectType": "ChatDto",
                                                        "_id": "66d8611a184d5f0a7c0675b7",
                                                        "name": "chat1"
                                                    },
                                                    {
                                                        "objectType": "ChatDto",
                                                        "_id": "66d85e3f4bed433708d5cb7f",
                                                        "name": "Chat 1"
                                                    }
                                                ]
                                            },
                                            "metaData": {
                                                "dataKey": "Chat",
                                                "modelName": "Chat",
                                                "dataType": "Array",
                                                "dataCount": 2
                                            },
                                            "paginationData": {
                                                "currentPage": 1,
                                                "limit": 20,
                                                "offset": 0,
                                                "itemCount": 2,
                                                "pageCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Read all existing chats",
                "description": "Read all created Chats. Remember about the pagination.\n\nNotice, that use of messages array is not advised and can be removed at some point in the future. For accessing messages of the Chat please use the /chat/:_id/message endpoint.\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | yes |\n| Authorization   | only for Players above 13 or with parents approval |\n| Search | yes | \n| Sort | yes |\n| Pagination | yes | ",
                "externalDocs": {
                    "url": "https://github.com/Alt-Org/Altzone-Server/wiki/3.-Data-fetching-(GET-requests)"
                }
            },
            "put": {
                "requestBody": {
                    "description": "Chat object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "409": {
                        "$ref": "#/components/responses/409"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Update a chat",
                "description": "Update the Chat, which _id is specified in the body.\n\nNotice that currently anybody is able to change any Chat.\n\n| Feature / Requirement | Is / Has | \n| -------- | -------- | \n| Authentication | yes | \n| Authorization   | only for Players above 13 or with parents approval | \n"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatCreate"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/201"
                                },
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Chat": {
                                                    "objectType": "ChatDto",
                                                    "_id": "66d86147184d5f0a7c0675ca",
                                                    "name": "chat2"
                                                }
                                            },
                                            "metaData": {
                                                "dataKey": "Chat",
                                                "modelName": "Chat",
                                                "dataType": "Object",
                                                "dataCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "409": {
                        "$ref": "#/components/responses/409"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Create a chat",
                "description": "Create a new Chat. The Chat is an object containing messages.\n\nNotice, that currently there is no restrictions on who can create a Chat.\n\nNotice that the Message objects are inner objects of Chat and can not be used enewhere else than in the Chat. There is also no separate collection for the Message in the DB.\n\n| Feature / Requirement | Is / Has | \n| -------- | -------- | \n| Authentication | yes | \n| Authorization   | only for Players above 13 or with parents approval | \n"
            }
        },
        "/chat/{_id}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200_one"
                                },
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Chat": {
                                                    "objectType": "ChatDto",
                                                    "_id": "66d8611a184d5f0a7c0675b7",
                                                    "name": "chat1"
                                                }
                                            },
                                            "metaData": {
                                                "dataKey": "Chat",
                                                "modelName": "Chat",
                                                "dataType": "Object",
                                                "dataCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Get Chat by _id",
                "description": "TEMPORARY DISABLED AND RETURNS 501\n\nRead Chat data by its _id field\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | yes |\n| Authorization   | only for Players above 13 or with parents approval |\n| Search | no | \n| Sort | no |\n| Pagination | no | "
            },
            "delete": {
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Delete Chat",
                "description": "TEMPORARY DISABLED AND RETURNS 501\n\nDelete Chat by its _id field. \n\nNotice that currently anybody can delete any Chat.\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | yes |\n| Authorization   | only for Players above 13 or with parents approval |"
            },
            "parameters": [
                {
                    "examples": {
                        "1": {
                            "value": "654fad58fd96261edd2e096f"
                        }
                    },
                    "name": "_id",
                    "description": "Chat _id field",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/chat/{_id}/messages": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200"
                                },
                                "examples": {
                                    "1": {
                                        "value": {
                                            "data": {
                                                "Chat": [
                                                    {
                                                        "id": 15,
                                                        "senderUsername": "user1",
                                                        "content": "message 15",
                                                        "feeling": 1
                                                    }
                                                ]
                                            },
                                            "metaData": {
                                                "dataKey": "Chat",
                                                "modelName": "Chat",
                                                "dataType": "Array",
                                                "dataCount": 1
                                            },
                                            "paginationData": {
                                                "currentPage": 1,
                                                "limit": 20,
                                                "offset": 0,
                                                "itemCount": 1,
                                                "pageCount": 1
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Read all messages from the Chat",
                "description": "Read all messages of specified Chat. Remember about the pagination\n\n| Feature / Requirement | Is / Has |\n| -------- | -------- |\n| Authentication | yes |\n| Authorization   | only for Players above 13 or with parents approval |\n| Search | yes | \n| Sort | yes |\n| Pagination | yes | ",
                "externalDocs": {
                    "url": "https://github.com/Alt-Org/Altzone-Server/wiki/3.-Data-fetching-(GET-requests)"
                }
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageCreate"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "409": {
                        "$ref": "#/components/responses/409"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "summary": "Create a message",
                "description": "Create a new Message. Message represent the object of message sent by a Player.\n\nNotice that currently there are no authorization required.\n\nNotice, that the messages does not have an usual _id field generated by data base. Instead the Photon id should be used.\n\nNotice that the messages is contained in the array of a Chat collection.\n\n| Feature / Requirement | Is / Has | \n| -------- | -------- | \n| Authentication | yes |\n| Authorization   | only for Players above 13 or with parents approval |\n"
            },
            "parameters": [
                {
                    "examples": {
                        "1": {
                            "value": "66912712d191c865ab53da8a"
                        }
                    },
                    "name": "_id",
                    "description": "Chat _id, with which the message is associated",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
            "200": {
                "description": "Found object(s). The response objects has three fields: data, metaData and paginationData.\nThe data has the requested data. It has an object or array, which key is a collection name.\n\nThe metaData contains additional data, which can be used for parsing the data object: \ndataKey - collection name, in data object,\ndataType - type of object in data object.\n\nThe paginationData is the data for pagination. Only on GET requests, when reading many objects\n",
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "metaData": {
                        "type": "object",
                        "properties": {
                            "dataKey": {
                                "description": "collection name",
                                "type": "string"
                            },
                            "modelName": {
                                "description": "model name, usually same as dataKey",
                                "type": "string"
                            },
                            "dataType": {
                                "description": "Object or Array",
                                "type": "string"
                            },
                            "dataCount": {
                                "description": "1 for Object dataType and length if Array",
                                "type": "integer"
                            }
                        }
                    },
                    "paginationData": {
                        "type": "object",
                        "properties": {
                            "currentPage": {
                                "description": "The current page number",
                                "type": "integer"
                            },
                            "limit": {
                                "description": "The limit of items per page",
                                "type": "integer"
                            },
                            "offset": {
                                "description": "The offset of items",
                                "type": "integer"
                            },
                            "itemCount": {
                                "description": "The total number of items",
                                "type": "integer"
                            },
                            "pageCount": {
                                "description": "The total number of pages",
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "201": {
                "description": "Created object. The response objects has two fields: data, metaData.\n\nThe data has the created object data. It has an object inside, which key is a collection name.\n\nThe metaData contains additional data, which can be used for parsing the data object: \ndataKey - collection name, in data object,\ndataType - type of object in the data object.",
                "required": [
                    "data",
                    "metaData"
                ],
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "metaData": {
                        "type": "object",
                        "properties": {
                            "dataKey": {
                                "description": "collection name",
                                "type": "string"
                            },
                            "modelName": {
                                "description": "model name",
                                "type": "string"
                            },
                            "dataType": {
                                "description": "always has Object value",
                                "type": "string"
                            },
                            "dataCount": {
                                "description": "always has 1 value",
                                "type": "integer"
                            }
                        }
                    }
                }
            },
...
            "200_one": {
                "description": "Found object. The response objects has two fields: data and metaData.\nThe data has the requested data. It has an object inside, which key is a collection name.\n\nThe metaData contains additional data, which can be used for parsing the data object: \ndataKey - collection name, in data object,\ndataType - type of object in data object.",
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "metaData": {
                        "type": "object",
                        "properties": {
                            "dataKey": {
                                "description": "collection name",
                                "type": "string"
                            },
                            "modelName": {
                                "description": "model name",
                                "type": "string"
                            },
                            "dataType": {
                                "description": "always has Object value",
                                "type": "string"
                            },
                            "dataCount": {
                                "description": "always has 1 value",
                                "type": "integer"
                            }
                        }
                    }
                }
            },
...
            "ChatCreate": {
                "description": "Create a new Chat",
                "type": "object",
                "properties": {
                    "name": {
                        "description": "The name of the chat",
                        "type": "string",
                        "example": "Chat 1"
                    }
                }
            },
            "ChatUpdate": {
                "description": "Update an existing Chat",
                "type": "object",
                "properties": {
                    "_id": {
                        "description": "The _id of the Chat, readonly",
                        "type": "string",
                        "example": "66912712d191c865ab53da8a"
                    },
                    "name": {
                        "description": "The name of the chat",
                        "type": "string",
                        "example": "Chat 1"
                    },
                    "messages": {
                        "description": "List of messages in the chat",
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "example": []
                    }
                }
            },
...
        },
        "responses": {
...
            "204": {
                "description": "Successful request, response with no body"
            },
            "400": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIConflictError"
                        },
                        "example": {
                            "statusCode": 400,
                            "message": [
                                "password must be a string"
                            ],
                            "error": "Bad Request"
                        }
                    }
                },
                "description": "Validation error. Some of the fields are not specified, have wrong data types or any over validation problem",
                "x-last-modified": 1718798714047
            },
            "401": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        },
                        "example": {
                            "statusCode": 401,
                            "message": "The access token is not provided. Please add `authorization` field with access token(in bearer token form): `Bearer access-token-here` to request header. The access token you can get from /auth endpoint",
                            "error": "Unauthorized"
                        }
                    }
                },
                "description": "Not authenticated. The Authorization header is not specified or the token is expired. [Here is more about auth](https://github.com/Alt-Org/Altzone-Server/wiki/2.-Authentication-and-authorization)"
            },
...
            "404": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        },
                        "example": {
                            "statusCode": 404,
                            "message": "Can not find any Chat instances",
                            "error": "Not Found"
                        }
                    }
                },
                "description": "Not found. No object(s) found"
            },
            "409": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIConflictError"
                        },
                        "example": {
                            "statusCode": 409,
                            "message": [
                                "Field 'username' with value 'user1' already exists"
                            ],
                            "error": "Conflict"
                        }
                    }
                },
                "description": "Some of the fields are not unique "
            },
            "500": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        },
                        "example": {
                            "statusCode": 500,
                            "message": "Internal server error"
                        }
                    }
                },
                "description": "Unexpected error happened. [Please create a new issue here](https://github.com/Alt-Org/Altzone-Server/issues) and specify the endpoint, HTTP method and description if u have any",
                "x-last-modified": 1718799573835
            },
...
        },
...
        "securitySchemes": {
            "JWTAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Example of header: \nAuthorization: Bearer access_token"
            }
        }
    },
    "security": [
        {
            "JWTAuth": []
        }
    ],
...
}